Renames vnode writer component to vnode publisher.
