+++ components/membership/group-policy-impl/build.gradle
@@ -24,7 +24,7 @@ dependencies {
 
     testImplementation "org.mockito.kotlin:mockito-kotlin:$mockitoKotlinVersion"
     testImplementation project(":components:membership:membership-static-memberlist-impl")
-    testImplementation project(":components:virtual-node:virtual-node-info-write-service")
+    testImplementation project(":components:virtual-node:virtual-node-publish-service")
     testImplementation project(":components:virtual-node:cpi-info-write-service")
     testImplementation project(":components:configuration:configuration-read-service")
     testImplementation project(":libs:messaging:messaging")
