+++ settings.gradle
@@ -151,7 +151,7 @@ include 'components:virtual-node:cpk-service-common'
 include 'components:virtual-node:sandbox-group-context-service'
 include 'components:virtual-node:virtual-node-common'
 include 'components:virtual-node:virtual-node-info-read-service'
-include 'components:virtual-node:virtual-node-info-write-service'
+include 'components:virtual-node:virtual-node-publish-service'
 include 'components:virtual-node:virtual-node-rpcops-service'
 include 'components:virtual-node:virtual-node-rpcops-service-impl'
 include 'libs:application:application-impl'
